backend/requirements.txt
fastapi
uvicorn
google-generativeai
spacy